Fix review issues: - fix typos; - move set operations into common;
